services/control-plane/requirements.txt
# Django
Django>=5.0,<5.1
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0
django-filter>=23.5,<24.0

# Database
psycopg[binary]>=3.1,<4.0
dj-database-url>=2.1,<3.0

# Authentication (OIDC/JWT)
PyJWT>=2.8,<3.0
cryptography>=41.0,<43.0

# API
pydantic>=2.5,<3.0
pydantic-settings>=2.1,<3.0

# Utilities
python-slugify>=8.0,<9.0
structlog>=23.2,<24.0
httpx>=0.26,<1.0

# Health checks
django-health-check>=3.17,<4.0

# Production
gunicorn>=21.2,<23.0
whitenoise>=6.6,<7.0

# Development
pytest>=7.4,<8.0
pytest-django>=4.7,<5.0
pytest-cov>=4.1,<5.0
factory-boy>=3.3,<4.0

# OpenAPI Documentation
drf-spectacular>=0.27,<1.0
hvac>=2.1,<3.0
